@@ -14,67 +14,68 @@
     "lint": "eslint src",
     "test": "jest --coverage",
     "clean": "rimraf dist",
-    "build": "npm run clean && npm run build:webpack",
+    "build": "npm run clean && npm run build:webpack_dev",
     "build:webpack": "cross-env NODE_ENV=production webpack --config webpack.config.prod.js",
     "build:webpack_dev": "cross-env NODE_ENV=development webpack --config webpack.config.prod.js",
     "bc": "npm run build && npm run build:copyall_win",
     "bu": "npm run build && npm run build:copyupd_win",
     "bv": "npm version patch",
     "build:copyall_win": "xcopy dist\\* ..\\WS\\public /E /Y /D && xcopy static\\* ..\\WS\\public /E /Y /D",
-    "build:copyupd_win": "copy dist\\bundle.js ..\\WS\\public\\ && copy dist\\style.css ..\\WS\\public\\"
+    "build:copyupd_win": "copy dist\\bundle.js ..\\WS\\public\\ && copy dist\\style.css ..\\WS\\public\\",
+    "build:copyall_win_mac": "npm run build && cp -r dist/* ../WS/public/ && cp -r static/* ../WS/public/",
+    "build:mac": "npm run build && mkdir -p ../WS/public && cp -r dist/* ../WS/public/ && cp -r static/* ../WS/public/",
+    "deploy:mac": "NODE_ENV=development npm run build:mac && cd ../WS && npm install && npm start"
   },
   "dependencies": {
     "ampersand-app": "^2.0.0",
-    "get-size": "^2.0.2",
-    "gsap": "^1.18.5",
-    "jquery": "^3.1.0",
-    "object-assign": "^4.1.0",
-    "react": "^15.2.0",
-    "react-dom": "^15.2.0",
-    "react-ga": "^2.1.1",
-    "react-imageloader": "^2.1.0",
-    "react-motion": "^0.4.4",
-    "react-router": "^2.5.2",
-    "socket.io": "^1.4.8",
-    "superagent": "^2.0.0",
-    "validator": "^5.4.0",
-    "x2js": "^2.0.1"
+    "framer-motion": "^10.18.0",
+    "get-size": "^2.0.3",
+    "gsap": "^3.12.5",
+    "jquery": "^3.7.1",
+    "object-assign": "^4.1.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-ga": "^3.3.1",
+    "react-image": "^4.1.0",
+    "react-motion": "^0.5.2",
+    "react-router-dom": "^6.22.2",
+    "socket.io-client": "^4.7.4",
+    "superagent": "^9.0.1",
+    "validator": "^13.11.0",
+    "x2js": "^3.4.4",
+    "xml-loader": "^1.2.1"
   },
   "devDependencies": {
-    "babel-core": "^6.10.4",
-    "babel-eslint": "^6.1.0",
-    "babel-jest": "^13.2.2",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-react-hmre": "^1.1.1",
-    "cross-env": "^1.0.8",
-    "css-loader": "^0.23.1",
-    "eslint": "^3.0.1",
-    "eslint-plugin-babel": "^3.3.0",
-    "eslint-plugin-react": "^5.2.2",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.14.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "imports-loader": "^0.6.5",
-    "jest-cli": "^13.2.3",
-    "node-sass": "^7.0.1",
-    "proxy-middleware": "^0.15.0",
-    "react-addons-test-utils": "^15.2.0",
-    "rimraf": "^2.5.3",
-    "sass-loader": "^4.0.0",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-hot-middleware": "^2.12.1"
+    "@babel/core": "^7.24.0",
+    "@babel/eslint-parser": "^7.23.10",
+    "@babel/preset-env": "^7.24.0",
+    "@babel/preset-react": "^7.23.3",
+    "babel-loader": "^9.1.3",
+    "cross-env": "^7.0.3",
+    "css-loader": "^6.10.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.33.2",
+    "express": "^4.18.3",
+    "express-http-proxy": "^2.0.0",
+    "file-loader": "^6.2.0",
+    "jest": "^29.7.0",
+    "mini-css-extract-plugin": "^2.8.1",
+    "rimraf": "^5.0.5",
+    "sass": "^1.71.1",
+    "sass-loader": "^14.1.1",
+    "style-loader": "^3.3.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.90.3",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-middleware": "^7.0.0",
+    "webpack-hot-middleware": "^2.26.1"
   },
   "jest": {
-    "unmockedModulePathPatterns": [
-      "<rootDir>/node_modules/react",
-      "<rootDir>/node_modules/react-dom",
-      "<rootDir>/node_modules/react-addons-test-utils"
-    ]
+    "testEnvironment": "jsdom",
+    "moduleNameMapper": {
+      "\\.(css|less|sass|scss)$": "<rootDir>/__mocks__/styleMock.js",
+      "\\.(gif|ttf|eot|svg)$": "<rootDir>/__mocks__/fileMock.js"
+    }
   }
 }
