@@ -1,29 +1,39 @@
-import React, {Component} from 'react'
-import PopUp from '../layouts/PopUp'
+import React, { Component } from "react";
+import { useParams } from "react-router-dom";
+import PopUp from "../layouts/PopUp";
+import PropTypes from "prop-types";
 
-export default class PopUp_page extends Component {
+// Create a wrapper component that uses hooks and passes params as props
+function PopUpPageWrapper(props) {
+  const params = useParams();
+  return <PopUp_page params={params} {...props} />;
+}
 
-	constructor (props) {
-		super(props)
-	}
+// Main component implementation
+class PopUp_page extends Component {
+  constructor(props) {
+    super(props);
+  }
 
-	render () {
-		const { pu_page } = this.props.params
-		const page_x = app.settings.ws_conf.pgs[pu_page]
+  render() {
+    const { pu_page } = this.props.params;
+    const page_x = app.settings.ws_conf.pgs[pu_page];
 
-		if (!pu_page || !page_x) return null
+    if (!pu_page || !page_x) return null;
 
-		// console.log(page_x)
+    // console.log(page_x)
 
-		return (
-			<PopUp pageTitle={page_x.pg_name}>
-				<div dangerouslySetInnerHTML={{__html: page_x.__cdata}} />
-			</PopUp>
-		)
-	}
-
+    return (
+      <PopUp pageTitle={page_x.pg_name}>
+        <div dangerouslySetInnerHTML={{ __html: page_x.__cdata }} />
+      </PopUp>
+    );
+  }
 }
 
 PopUp_page.propTypes = {
-	params: React.PropTypes.any
-}
+  params: PropTypes.any,
+};
+
+// Export the wrapper as the default
+export default PopUpPageWrapper;
