@@ -1,49 +1,52 @@
-var path = require('path')
-var webpack = require('webpack')
+const path = require("path");
+const webpack = require("webpack");
 
 module.exports = {
-	devtool: 'cheap-module-eval-source-map',
-	// context: path.join(__dirname, 'static'),
-	entry: [
-		'eventsource-polyfill', // necessary for hot reloading with IE
-		'webpack-hot-middleware/client',
-		'./src/app'
-	],
-	output: {
-		path: path.join(__dirname, 'dist'),
-		filename: 'bundle.js',
-		// publicPath: path.join(__dirname, 'static'),
-		// publicPath: __dirname + '/static/'
-		publicPath: '/'
-	},
-	plugins: [
-		new webpack.HotModuleReplacementPlugin(),
-		new webpack.NoErrorsPlugin()
-	],
-	module: {
-		loaders: [
-			{
-				test: /\.(ico|css|gif|png|html|jpg|xml|svg)$/,
-					// loader: 'url?limit=10000',
-				loader: 'file?name=[path][name].[ext]&context=./static',
-			},
-			{
-				test: /\.jsx?/,
-				loaders: ['babel'],
-				include: path.join(__dirname, 'src')
-			},
-			{
-				test: /\.scss$/,
-				loaders: ['style', 'css', 'sass']
-			},
-			{
-				test: /\.css$/,
-				loaders: ['style', 'css']
-			},
-			{
-				test: /(flickity|fizzy-ui-utils|get-size|unipointer|imagesloaded)/,
-				loader: 'imports?define=>false&this=>window'
-			},
-		]
-	},
-}
+  mode: "development",
+  // context: path.join(__dirname, 'static'),
+  entry: ["webpack-hot-middleware/client", "./src/index.jsx"],
+  output: {
+    path: path.join(__dirname, "dist"),
+    filename: "bundle.js",
+    // publicPath: path.join(__dirname, 'static'),
+    publicPath: "/static/",
+  },
+  resolve: {
+    extensions: [".js", ".jsx"],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: "babel-loader",
+        },
+      },
+      {
+        test: /\.s[ac]ss$/i,
+        use: ["style-loader", "css-loader", "sass-loader"],
+      },
+      {
+        test: /\.css$/,
+        use: ["style-loader", "css-loader"],
+      },
+      {
+        test: /\.(png|jpg|gif|svg|eot|ttf|woff|woff2)$/,
+        type: "asset",
+        parser: {
+          dataUrlCondition: {
+            maxSize: 8192,
+          },
+        },
+      },
+    ],
+  },
+  plugins: [
+    new webpack.HotModuleReplacementPlugin(),
+    new webpack.DefinePlugin({
+      "process.env.NODE_ENV": JSON.stringify("development"),
+    }),
+  ],
+  devtool: "eval-source-map",
+};
